# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是软件杯B1赛题",
        "title": "龙芯服务 API",
        "termsOfService": "http://swagger.io/terms/ 使用条款",
        "contact": {
            "name": "LoveSong.",
            "url": "无",
            "email": "3130250166@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "39.99.139.249:8080",
    "paths": {
...
        "/files/compress": {
            "post": {
                "description": "压缩文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "压缩文件",
                "parameters": [
                    {
                        "description": "request",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/files.FileCompress"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/content": {
            "post": {
                "description": "查看文件属性和查看文件内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "查看文件属性和查看文件内容",
                "parameters": [
                    {
                        "description": "request",
                        "name": "content",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/files.FileContentReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/files.FileInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/create": {
            "post": {
                "description": "创建文件 IsSymLink软链接 IsLink硬链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "创建文件",
                "parameters": [
                    {
                        "description": "request",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/files.CreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/decompress": {
            "post": {
                "description": "解压文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "解压文件",
                "parameters": [
                    {
                        "description": "request",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/files.FileDecompress"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/download": {
            "get": {
                "description": "下载文件指明路径",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "下载文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "path",
                        "name": "path",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/files/mode": {
            "post": {
                "description": "修改文件权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "修改文件权限",
                "parameters": [
                    {
                        "description": "request",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/files.ChmodReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/recover": {
            "post": {
                "description": "恢复回收站(单个文件)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "恢复回收站(单个文件)",
                "parameters": [
                    {
                        "description": "request",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.RecoverReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/recycle/search": {
            "post": {
                "description": "查询回收站(分页)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "查询回收站(分页)",
                "parameters": [
                    {
                        "description": "request",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.PageInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/remove": {
            "post": {
                "description": "删除文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "删除文件",
                "parameters": [
                    {
                        "description": "request",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/files.RemoveReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/rename": {
            "post": {
                "description": "修改文件名字",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "修改文件名字",
                "parameters": [
                    {
                        "description": "request",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/files.FileRenameReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/search": {
            "post": {
                "description": "获取该路径的文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "获取该路径的文件",
                "parameters": [
                    {
                        "description": "request",
                        "name": "FileOption",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/files.FileOption"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/files.FileInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/size": {
            "post": {
                "description": "查看文件大小",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "查看文件大小",
                "parameters": [
                    {
                        "description": "request",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/files.DirSizeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/files.DirSizeRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/upload": {
            "post": {
                "description": "上传文件指明路径 file 上传的文件 path存放的路径 cover 是否覆盖 用 postman的 body form-data",
                "tags": [
                    "File"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "request",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "files.ChmodReq": {
            "type": "object",
            "required": [
                "mode",
                "path"
            ],
            "properties": {
                "mode": {
                    "description": "十进制",
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "files.CreateReq": {
            "type": "object",
            "required": [
                "path"
            ],
            "properties": {
                "isDir": {
                    "type": "boolean"
                },
                "isLink": {
                    "type": "boolean"
                },
                "isSymLink": {
                    "type": "boolean"
                },
                "linkPath": {
                    "type": "string"
                },
                "mode": {
                    "description": "十进制",
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "files.DirSizeReq": {
            "type": "object",
            "required": [
                "path"
            ],
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
        "files.DirSizeRes": {
            "type": "object",
            "properties": {
                "size": {
                    "type": "number"
                }
            }
        },
        "files.FileCompress": {
            "type": "object",
            "required": [
                "dst",
                "files",
                "name",
                "type"
            ],
            "properties": {
                "dst": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "replace": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "files.FileContentReq": {
            "type": "object",
            "required": [
                "path"
            ],
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
        "files.FileDecompress": {
            "type": "object",
            "required": [
                "dst",
                "path",
                "type"
            ],
            "properties": {
                "dst": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "files.FileInfo": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "extension": {
                    "type": "string"
                },
                "favoriteID": {
                    "type": "integer"
                },
                "gid": {
                    "type": "string"
                },
                "group": {
                    "type": "string"
                },
                "isDir": {
                    "type": "boolean"
                },
                "isHidden": {
                    "type": "boolean"
                },
                "isSymlink": {
                    "type": "boolean"
                },
                "itemTotal": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/files.FileInfo"
                    }
                },
                "linkPath": {
                    "type": "string"
                },
                "mimeType": {
                    "type": "string"
                },
                "modTime": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "files.FileOption": {
            "type": "object",
            "required": [
                "path"
            ],
            "properties": {
                "containSub": {
                    "type": "boolean"
                },
                "dir": {
                    "type": "boolean"
                },
                "expand": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "search": {
                    "type": "string"
                },
                "showHidden": {
                    "type": "boolean"
                },
                "sortBy": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "string"
                }
            }
        },
        "files.FileRenameReq": {
            "type": "object",
            "required": [
                "newName",
                "oldName",
                "path"
            ],
            "properties": {
                "newName": {
                    "type": "string"
                },
                "oldName": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "files.RemoveReq": {
            "type": "object",
            "required": [
                "path"
            ],
            "properties": {
                "isDir": {
                    "type": "boolean"
                },
                "path": {
                    "type": "string"
                },
                "realDel": {
                    "type": "boolean"
                }
            }
        },
...
        "model.PageInfo": {
            "type": "object",
            "required": [
                "page",
                "pageSize"
            ],
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                }
            }
        },
        "model.RecoverReq": {
            "type": "object",
            "required": [
                "from",
                "name",
                "rName"
            ],
            "properties": {
                "from": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rName": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}